new CONFVER and renamed a lot from pwm to gpio; compile with esp too

+++ arduino/sketchbook/libraries/i2cgpio/examples/test_gpio/test_gpio.ino
@@ -29,7 +29,7 @@ void setup() {
 
 void loop() {
 int steps=400;
-int position=0;
+int16_t position=0;
   Log.notice(F("digitalWrite" CR));  
   gpio.digitalWrite(1,HIGH);
   gpio.digitalWrite(2,HIGH);
